RT_application/requirements.txt
librosa==0.10.1
flask==3.0.2
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.15.0